Move UIWindow Extension to new file
