+++ AnotherFeedlyApp/AppDelegate.swift
@@ -19,16 +19,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 }
 
-extension UIWindow {
-    static func configure() -> UIWindow {
-        let screenFrame = UIScreen.main.bounds
-        let window = UIWindow(frame: screenFrame)
-        window.backgroundColor = .blue
-        window.makeKeyAndVisible()
-        return window
-    }
-}
-
 @objc(TestingAppDelegate)
 class TestingAppDelegate: UIResponder, UIApplicationDelegate {
